+++ react/__mocks__/vtex.render-runtime.js
@@ -5,3 +5,7 @@ export function withRuntimeContext(Comp) {
 }
 
 export const Link = ({ children }) => <a>{children}</a>
+
+export const NoSSR = ({ children }) => (
+  <div className="NoSSR-mock">{children}</div>
+)